requirements.txt
flask>=2.2.0
requests>=2.28.0
python-dotenv>=0.20.0
Flask-Cors>=3.0.10
pytz>=2022.1